# openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Product": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "priceCents": {
            "title": "Pricecents",
            "type": "integer"
          },
          "stock": {
            "title": "Stock",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "priceCents",
          "stock",
          "available"
        ],
        "title": "Product",
        "type": "object"
      },
      "Reservation": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ReservedLine"
            },
            "title": "Lines",
            "type": "array"
          },
          "orderRef": {
            "title": "Orderref",
            "type": "string"
          },
          "totalCents": {
            "title": "Totalcents",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "orderRef",
          "lines",
          "totalCents"
        ],
        "title": "Reservation",
        "type": "object"
      },
      "ReservationItem": {
        "properties": {
          "productId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Productid",
            "type": "string"
          },
          "quantity": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "productId",
          "quantity"
        ],
        "title": "ReservationItem",
        "type": "object"
      },
      "ReservationRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReservationItem"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "orderRef": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Orderref",
            "type": "string"
          }
        },
        "required": [
          "orderRef",
          "items"
        ],
        "title": "ReservationRequest",
        "type": "object"
      },
      "ReservedLine": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "priceCents": {
            "title": "Pricecents",
            "type": "integer"
          },
          "productId": {
            "title": "Productid",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "productId",
          "name",
          "quantity",
          "priceCents"
        ],
        "title": "ReservedLine",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "wmd catalog-service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/products": {
      "get": {
        "operationId": "list_products_v1_products_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  },
                  "title": "Response List Products V1 Products Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Products"
      }
    },
    "/v1/products/{product_id}": {
      "get": {
        "operationId": "get_product_v1_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Product"
      }
    },
    "/v1/reservations": {
      "post": {
        "description": "Take stock for every item or none: all-or-nothing in one transaction.",
        "operationId": "reserve_v1_reservations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reserve"
      }
    }
  }
}